+++ README.md
@@ -40,7 +40,7 @@ Each protocol should be placed in directory `base_dir/protocol`.
 `protocol` can be single directory or multiple subdirectories, outer directories are used as namespace for generated messages, e.g. "my_company/core" or "my_company/the_product/protocol".
 
 Message generator usage:
-```
+```bash
 python3 messgen-generate.py --basedir <base_dir> --protocol <protocol> --lang <lang> --outdir <out_dir> [--options key1=value1,key2=value2,...]
 ```
 
@@ -50,7 +50,7 @@ Generated messages placed in `out_dir` directory.
 
 Example for C++ messages generation:
 
-```
+```bash
 python3 messgen-generate.py --basedir ./base_dir --protocol my_namespace/my_protocol --lang cpp --outdir out/cpp --options cpp_standard=20
 ```
 
@@ -58,7 +58,7 @@ python3 messgen-generate.py --basedir ./base_dir --protocol my_namespace/my_prot
 
 Example for JS messages generation:
 
-```
+```bash
 python3 messgen-generate.py --basedir ./base_dir --protocol my_namespace/my_protocol --lang json --outdir out/json
 ```
 This command will generate json schema containing full protocol description.
