+++ port/js/benchmarks/deserialize-variant.bench.js
@@ -1,11 +1,11 @@
 import { bench } from 'vitest'
 
-import {Struct} from '../src/messgen.js';
 import {Buffer} from './deserialize-variant/messgen-old.js';
 import { Buffer as  BufferFromEntries} from './deserialize-variant/messgen-fromEntries.js';
 import { Buffer as  BufferPreBuild, Struct as StructPreBuild} from './deserialize-variant/messgen-pre-build.js';
 
 import { bench, describe } from 'vitest'
+import {Struct} from "../src/Struct.ts";
 
 let srcStruct = new Struct({
   id: 2,
